DROID-2123 compose icon
